[refactor] #47 - GroupData 위치 이동 및 불필요한 파일 삭제
